+++ src/main/java/edu/cornell/kfs/tax/businessobject/options/FormType1099ValuesFinder.java
@@ -8,7 +8,7 @@ import org.kuali.kfs.krad.keyvalues.KeyValuesBase;
 import org.kuali.rice.core.api.util.ConcreteKeyValue;
 import org.kuali.rice.core.api.util.KeyValue;
 
-import edu.cornell.kfs.tax.CUTaxConstants.FormTypes1099;
+import edu.cornell.kfs.tax.FormTypes1099;
 
 public class FormType1099ValuesFinder extends KeyValuesBase {
 	private static final long serialVersionUID = 2880415540151030542L;
